schematic v1.0 for review
First pass at schematic for review. The original design called for two
opamps - a summing LMP7731 and an inverting TLV2460. Unfortunately the
TLV2460 has been discontinued, so I took the opportunity to update the
design for the OPA322. This appears to offer better noise
characteristics while retaining rail-to-rail swing @ >20mA output,
required to drive 1.0V into 50ohms.

Major component placement is complete on the PCB & front panel sketches
are done.